+++ value.go
@@ -82,12 +82,14 @@ func NewValue(iso *Isolate, val interface{}) (*Value, error) {
 			rtnVal = &Value{
 				ptr: C.NewValueBigInt(iso.ptr, C.int64_t(v.Int64())),
 			}
+			break
 		}
 
 		if v.IsUint64() {
 			rtnVal = &Value{
 				ptr: C.NewValueBigIntFromUnsigned(iso.ptr, C.uint64_t(v.Uint64())),
 			}
+			break
 		}
 
 		var sign, count int
@@ -517,7 +519,16 @@ func (v *Value) AsObject() (*Object, error) {
 }
 
 func (v *Value) finalizer() {
-	C.ValueDispose(v.ptr)
+	C.ValueFree(v.ptr)
 	v.ptr = nil
 	runtime.SetFinalizer(v, nil)
+}
+
+// MarshalJSON implements the json.Marshaler interface.
+func (v *Value) MarshalJSON() ([]byte, error) {
+	jsonStr, err := JSONStringify(nil, v)
+	if err != nil {
+		return nil, err
+	}
+	return []byte(jsonStr), nil
 }